backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1  # AsyncIO MongoDB driver
newspaper3k==0.2.8  # For news scraping
transformers==4.35.2  # For sentiment analysis
torch==2.1.1  # Required for transformers
python-dotenv==1.0.0
pydantic==2.5.1
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose==3.3.0  # For JWT tokens
passlib==1.7.4  # For password hashing
python-multipart==0.0.6  # For form data
loguru==0.7.2  # For better logging
httpx==0.25.2  # For async HTTP requests